+++ src/Dropdown/Dropdown.tsx
@@ -11,7 +11,7 @@ import styled from 'styled-components'
 
 import { theme } from '../theme'
 import { Icon } from '../Icon'
-import { Field, CommonFieldTypes } from '../Field'
+import { Field, CommonFieldTypes } from '../fields/Field'
 import { useUniqueId } from '../utils/id'
 
 export type DropdownItem = {
