+++ src/index.ts
@@ -30,6 +30,6 @@ export * from './Textarea'
 export * from './TextInput'
 export * from './theme'
 export * from './Toggle'
-export * from './Fieldset'
+export * from './fields/Fieldset'
 export * from './fontStyle'
 export { MarginProps, PaddingProps } from './utils/space'
